main/livekit-server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.9"

dependencies = [
    "livekit-agents[openai,turn-detector,silero,cartesia,deepgram]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-groq",
    "livekit.plugins.deepgram",
    "livekit-plugins-elevenlabs",
    "livekit-plugins-inworld",
    "python-dotenv",
    "paho-mqtt",
    "aiohttp",
    "pydub",
    "edge-tts",
    "gtts",
    "pyttsx3",
    "mem0ai>=1.0.1",
    "jinja2",
    "pytz",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.0",
    "websockets>=11.0",
    "torchaudio>=2.8.0",
    "silero-vad==6.2.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"
[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
